Move pre fork2 producers to another file
