+++ crates/sequencer/src/lib.rs
@@ -21,6 +21,7 @@ use tokio::sync::broadcast;
 use tokio::sync::mpsc::unbounded_channel;
 
 mod commitment;
+mod compat;
 pub mod db_migrations;
 mod db_provider;
 mod deposit_data_mempool;
